# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hubbub team command-center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{slug}/items/{itemNumber}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on an item
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: path
          name: itemNumber
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Post a comment
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: path
          name: itemNumber
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /projects/{slug}/items/{itemNumber}/comments/{commentId}:
    patch:
      operationId: updateComment
      tags: [comments]
      summary: Edit comment
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: path
          name: itemNumber
          required: true
          schema:
            type: integer
        - in: path
          name: commentId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete comment
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: path
          name: itemNumber
          required: true
          schema:
            type: integer
        - in: path
          name: commentId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Time Entries ───────────────────────────────────────────────────────────
...
  /projects/{slug}/costs:
    get:
      operationId: listCostEntries
      tags: [costs]
      summary: List cost entries for a project
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cost entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CostEntry"
    post:
      operationId: createCostEntry
      tags: [costs]
      summary: Create cost entry
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CostEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostEntry"
  /projects/{slug}/costs/{costId}:
    delete:
      operationId: deleteCostEntry
      tags: [costs]
      summary: Delete cost entry
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: path
          name: costId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Commits ────────────────────────────────────────────────────────────────
...
  /projects/{slug}/docs:
    get:
      operationId: listDocs
      tags: [docs]
      summary: List docs for a project
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Docs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doc"
    post:
      operationId: createDoc
      tags: [docs]
      summary: Create doc
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doc"
  /projects/{slug}/docs/{docSlug}:
    get:
      operationId: getDoc
      tags: [docs]
      summary: Get doc by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: path
          name: docSlug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Doc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doc"
        "404":
          description: Not found
    patch:
      operationId: updateDoc
      tags: [docs]
      summary: Update doc
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: path
          name: docSlug
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doc"
    delete:
      operationId: deleteDoc
      tags: [docs]
      summary: Delete doc
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: path
          name: docSlug
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Search ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, displayName, role, active]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: ["string", "null"]
        displayName:
          type: string
        username:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, member]
        hourlyRateCents:
          type: ["integer", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Comment:
      type: object
      required: [id, itemId, authorId, body, createdAt]
      properties:
        id:
          type: integer
        itemId:
          type: integer
        authorId:
          type: string
        author:
          $ref: "#/components/schemas/User"
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CommentInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
    CommentUpdate:
      type: object
      required: [body]
      properties:
        body:
          type: string

    # ── Commits ───────────────────────────────────────────────────────────
...
    CostEntry:
      type: object
      required: [id, projectId, category, amountCents, currency, recurring, incurredOn]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        scopeId:
          type: ["integer", "null"]
        category:
          type: string
          enum: [labor, hosting, saas, contractor, ai, other]
        vendor:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        amountCents:
          type: integer
        currency:
          type: string
        recurring:
          type: boolean
        incurredOn:
          type: string
          format: date
    CostEntryInput:
      type: object
      required: [category, amountCents, incurredOn]
      properties:
        scopeId:
          type: ["integer", "null"]
        category:
          type: string
          enum: [labor, hosting, saas, contractor, ai, other]
        vendor:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        amountCents:
          type: integer
        currency:
          type: string
        recurring:
          type: boolean
        incurredOn:
          type: string
          format: date

    # ── Presence ──────────────────────────────────────────────────────────
...
    Doc:
      type: object
      required: [id, projectId, title, slug, body, pinned, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        slug:
          type: string
        body:
          type: string
        pinned:
          type: boolean
        createdById:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DocInput:
      type: object
      required: [title, slug, body]
      properties:
        title:
          type: string
        slug:
          type: string
        body:
          type: string
        pinned:
          type: boolean
    DocUpdate:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        pinned:
          type: boolean

    # ── Search ────────────────────────────────────────────────────────────
...