# lib/api-zod/src/generated/types/docUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hubbub team command-center API
 * OpenAPI spec version: 0.1.0
 */

export interface DocUpdate {
  title?: string;
  body?: string;
  pinned?: boolean;
}